+++ app/src/main/java/com/example/cryptoapp/data/network/ApiFactory.kt
@@ -6,7 +6,6 @@ import retrofit2.converter.gson.GsonConverterFactory
 object ApiFactory {
 
     private const val BASE_URL = "https://min-api.cryptocompare.com/data/"
-    const val BASE_IMAGE_URL = "https://cryptocompare.com"
 
     private val retrofit = Retrofit.Builder()
         .addConverterFactory(GsonConverterFactory.create())
